requirements.txt
pandas==2.2.2
numpy==1.26.4
sentence-transformers==3.4.1
transformers==4.48.2
kaleido==0.2.1
datasets==3.2.0
regex==2024.11.6
requests==2.32.3
matplotlib==3.10.0
plotly==5.24.1
nltk==3.9.1
wordcloud==1.9.4
text-unidecode==1.3
openai==1.61.1
matplotlib==3.10.0
pydantic==2.10.6
pydantic_core==2.27.2
fastapi
uvicorn
python-dotenv
